pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling"

[project]
name = "robot_sim"
version = "0.1.0"
description = "Unified Robot Simulation Platform for Isaac Lab and MuJoCo"
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = { file = "LICENSE" }
authors = [{ name = "Yongce Liu", email = "yongce.liu@outlook.com" }]
keywords = ["robotics", "simulation", "isaac-lab", "mujoco", "unitree"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "numpy>=1.20.0",
    "pyzmq>=25.0.0",  # For model communication
    # Add simulator-specific dependencies as optional
    # "mujoco>=2.3.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "ruff", "mypy>=0.990"]

[tool.uv]
package = true
python-preference = "managed"

[tool.uv.build-backend]
module-name = "robot_sim"
module-root = ""

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "WARNING"
addopts = "-s"

[tool.ruff]
line-length = 120

[tool.ruff.lint.extend-per-file-ignores]
# Also ignore `E402` in all `__init__.py` files.
"__init__.py" = ["E402", "F401", "F403", "E501"]

[tool.ruff.lint]
extend-select = ["UP"]
ignore = ["F722"]


# mypy type checking configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

# [project.urls]
# Homepage = "https://github.com/yourusername/robot_sim"
# Repository = "https://github.com/yourusername/robot_sim"

# # pytest configuration
# [tool.pytest.ini_options]
# testpaths = ["tests"]
# python_files = ["test_*.py"]
# python_classes = ["Test*"]
# python_functions = ["test_*"]

